Finished Registration in DI
